pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "biomancy"
version = "0.0.1"
authors = [
    { name = "Aleksandr Fedorov", email = "anfedrov@hse.ru" },
]
description = "TODO"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pybedtools ~= 0.9.0",
    "biopython ~= 1.81",
#    "torch ~= 2.0.0",
    "intervaltree ~= 3.1.0",
    "pyBigWig ~= 0.3.22",
    "joblib ~= 1.2.0",
    "scikit-learn ~= 1.2.2"
]

[project.optional-dependencies]
develop = [
    'pytest ~= 7.2.2',
    'wemake-python-styleguide ~= 0.17.0',
]

[project.urls]
"Homepage" = "https://github.com/hse-bioinflab/biomancy"
"Bug Tracker" = "https://github.com/hse-bioinflab/biomancy/issues"